slight refactor. Get transparent blocks to work
